Merge Db and Auth modules to core.
Since I have never ever used this as just a dumb router, but as a
framework for implementing data APIs, it seemed silly to continue
updating the Db and Auth modules separately. Bringing them back into
the core.

Also updating to PHP 5.4 because I tend to use [] to define arrays and
that messed me up when I deployed to a system that was not yet updated
past 5.3.