Controller->Authorize change
If an array is defined for the controller’s skipAuthentication array
and it contains either ‘*’ or the method name being called, no
authentication will be done.

If requireAuthentication is empty, assume that any method in this
controller should require authentication as an AuthInterface was
defined for the class.

If requireAuthentication array contains either ‘*’ or the name of the
method being called, authenticate the user.