Refactor Db/Mapper::translateWhere
Do not filter on a non-existent column. If an incorrect column is
included, the entire where clause should be rejected and set to blank.

This is a bad idea for a couple reasons, so I need to go back and
change the way that db filtering works in Mapper objects, but this is a
BC-breaking change, so I will attend to that at a later date.