REST Auth handler
Added a generic auth handler. When given a class implementing
Fluxoft\Rebar\Auth\AuthInterface, it will return the following:
GET - If the user is authenticated, an object containing “auth: true”
and “userID: [current user’s ID]”. If not authenticated, “auth” is
false and the userID is null.
POST - Given a credentials object containing “username” and “password”
(and optionally, a true/false value for “remember” which defaults to
false), a Login is attempted and the return object contains an “auth”
value of true if successful or false if unsuccessful. If successful,
the logged-in user’s ID is returned. Otherwise, the UserID is set to 0.
DELETE - Calls the Logout method on the given AuthInterface. Returns
“auth: false”