Router/Controller simplification
Setup() and Cleanup() will now receive the values in the Router’s
$config[‘setupArgs’] and $config[‘cleanupArgs’] respectively as method
parameters, rather than passing the entire array to the method. This is
just to make it simpler to define a Setup method (for instance) that
receives (for instance) a Container object, changing this:

```
public function Setup(array $params) {
    $this->container = $params[0];
}
```

to this:

```
public function Setup(Container $container) {
    $this-container = $container;
}
```

This should allow for more readable and maintainable code.