Remove Setup and Cleanup from Controller class
Because the intention is to allow for defining custom Setup() and
Cleanup() methods that might take 0 or more parameters, defining these
methods in the abstract controller will cause a strict standards
conflict if the derived class defines more parameters.

Only call Setup and Cleanup if they exist in the class.