Rest rework.
Move cross-origin header stuff into the Authorize method of the base
controller class. Cleaned up some unused methods in that class.

Moved preflight checking from the Rest controller to be part of the
Authorization process, since that actually makes more sense.

Greatly simplified the RepositoryInterface by just passing a Request
object and a params array to the methods there. This should make the
process of building custom repositories that do no use DataRepository
much simpler.

Improved exception handling in the Router.

fixes #34
This breaks BC for previous implementations of the Rest module, so this
stuff will all be in the 0.19.0 release.