Auth rework
Reorganization of code and preparation for introduction of JWT auth
type. This will break backwards compatibility for anyone who was using
Auth to handle restricting access to an application.

Login and GetAuthenticatedUser methods of AuthInterface now return an
instance of \Fluxoft\Rebar\Auth\Reply, which contains the User that was
originally returned.

Also, Basic authentication now uses a User model rather than an array
of usernames and passwords, so those will need to be redefined.