Router rework
closes #45

Big changes to the Router class are going to break everything, so this
will be part of a new minor version. Upgrade at your own risk!

Also includes an Error\Handler class that can be used to set default
error and exception handlers, passing in a class implementing
NotifierInterface that can do some kind of error handling and notifying
(the included BasicNotifier just spits the exception out to the screen,
but a custom Notifier might also log the error somewhere, email
someone, etc).