DataRepository changes
These changes will be seamless to existing implementations and will not
require any changes. They will allow a child class to access and change
the data that is sent for a PUT or POST request before it is processed
by calling DataRepository::get[Post|Put]Data, changing the returned
array, and then calling DataRepository::set[Post|Put]Data. After that,
the child class can just call parent::Put or parent::Post and the new
data set will be processed. Useful for preventing some properties from
being modified, or for removing some extra properties that might be
sent to an endpoint but that must be handled differently.

closes #48