Return 409 on POST if object already exists
If a unique constraint is violated, return a more specific error to
make it easier for the client to understand how to deal with it.
Before, this error would have fallen through to the generic 500
Database Error message, which is misleading.

Consider writing custom exceptions to be thrown at the mapper level for
these.