Do not save non-db properties
If a property is defined in Model::properties but not in
Model::propertyDbMap, it should not be used to build the SQL necessary
to save to the database. If a property is in one and not the other, it
is meant to be a read-only property, possibly the result of a derived
or aggregate query.