data mapper overhaul
This one breaks backward compatibility because of several changes to
the Db\Mapper and how SQL statements are constructed.

The Mapper can now have a custom SQL statement set that will return an
aliased list of properties, to allow for including derived fields in a
database model when retrieving models.

Filtering is now done by passing in a set of properties to be used as
filters and those are tacked on to the base selectSql string in either
a WHERE or HAVING clause, depending on whether the properties exist in
the `$propertyDbMap` or in `$properties`. (Only fields that can be
mapped to database columns should be included in `$propertyDbMap`.
Derived fields should be specified in properties).

Models can now include validation methods that will be called when
saving the object. A method named “Validate{PropertyName}” will be
called for each property, if defined. The method should return true if
validation succeeds, or an error message. If any validation method
fails, an `InvalidModelException` will be thrown rather than saving.
Validation errors can be accessed from the Model’s
`GetValidationErrors()` method.

This will be for 0.18.0