Auth/Db/User fix
Because of the change to how Model properties are set in the
constructor, the Auth\Db\User::setPassword function is called when
constructing a user, so the password was being re-hashed. Only generate
a hash if the password property was already set to something.
Otherwise, assume it is a constructor call and set it to whatever is
passed to this function.