fixes
Had to change the Environment test, because now apparently phpunit
knows how to simulate getallheaders() or it is included in a different
way in php7.1
Also removed the requirement to test on php7.0, as certain libraries
are now no longer supporting it. Also changed the requirement for rebar
to use php7.1 at minimum.