DB\Mapper refactor
Added a new Filter object to allow for more advanced filtering in the WHERE and HAVING clauses used to pull data. Now supports =,<,>,<>,<=,>=,IN, and BETWEEN as comparison operators. Still need to add some checks in the Filter object itself to throw exceptions in certain cases like IN or BETWEEN where the Value property must be an array to work (and in the case of BETWEEN must be an array with just two values)